requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
sqlalchemy
psycopg2-binary
pandas
python-dotenv
httpx
